+++ custom_components/zambretti_sager/const.py
@@ -2,7 +2,7 @@ from typing import Final
 
 DOMAIN = "zambretti_sager"
 
-VERSION = "1.9.5"
+VERSION = "1.9.10"
 
 # Frontend
 URL_BASE: Final[str] = "/zambretti_sager_card"